Добавит модель для отрисовки фотографий других пользователей
